chore: apply naming & styling conventions
