+++ src/components/chats/components/CreateNewChat/CreateSharedChat.tsx
@@ -1,23 +1,25 @@
 import { Button } from '@chakra-ui/react'
 import { DocumentIcon } from '@heroicons/react/24/outline'
-import { useCreatePrivateChat } from '../../chatHooks'
+import { useCreateSharedChat } from '../../chatHooks'
 import { NewChatProps } from './createNewChatTypes'
 import { useCanCreateChat } from './useCanCreateChat'
 
 export const CreateSharedChat = ({
   postId,
   chatId,
-  resetTextArea,
+  onSuccess,
 }: NewChatProps) => {
-  const { mutate: createChat } = useCreatePrivateChat()
+  const { mutate: createChat } = useCreateSharedChat()
   const canCreateChat = useCanCreateChat(postId, chatId)
 
   const handleCreateChat = () => {
-    createChat({
-      onSuccess: () => {
-        resetTextArea()
+    if (!postId) return
+    createChat(
+      { postId },
+      {
+        onSuccess,
       },
-    })
+    )
   }
 
   return (
@@ -26,8 +28,6 @@ export const CreateSharedChat = ({
       variant="outline"
       size="xs"
       className="mr-2"
-      // Revisit this: Is it the right behavior?
-      // https://www.notion.so/joiahq/Revisit-the-behaviour-of-the-disabled-new-chat-button-in-the-chatbox-f44da7d966754e439bd96e74bf564a60
       disabled={!canCreateChat}
     >
       <DocumentIcon className="-ml-0.5 mr-1 h-3 w-3" /> New chat
