+++ src/components/chats/components/CreateNewChat/createNewChatTypes.ts
@@ -1,7 +1,7 @@
 export interface NewChatProps {
   postId?: string
   chatId?: string
-  resetTextArea: () => void
+  onSuccess: () => void
 }
 
 export interface CreateNewChatProps extends NewChatProps {
