+++ src/components/sidebar/components/sidebarDesktop/SidebarDesktopDeleteChat/SidebarDesktopDeleteChat.tsx
@@ -1,12 +1,11 @@
-import { FC } from 'react'
 import { DeletePrivateChat } from './DeletePrivateChat'
 import { DeleteSharedChat } from './DeleteSharedChat'
-import { SidebarDesktopDeleteChatProps } from './types'
+import { SidebarDesktopDeleteChatProps } from './sideBarDesktopDeleteChatTypes'
 
-export const SidebarDesktopDeleteChat: FC<SidebarDesktopDeleteChatProps> = ({
+export const SidebarDesktopDeleteChat = ({
   isPrivate,
   ...props
-}) => {
+}: SidebarDesktopDeleteChatProps) => {
   if (isPrivate) return <DeletePrivateChat {...props} />
 
   return <DeleteSharedChat {...props} />
